+++ packages/core/base/src/index.ts
@@ -2,4 +2,5 @@ export * from './adapter.js';
 export * from './errors.js';
 export * from './signer.js';
 export * from './standard.js';
+export * from './transaction.js';
 export * from './types.js';
